[dxso] Migrate RegisterLinkerSlot to dxso_util
