+++ src/dxso/dxso_util.h
@@ -3,6 +3,7 @@
 #include <cstdint>
 
 #include "dxso_common.h"
+#include "dxso_decoder.h"
 
 namespace dxvk {
 
@@ -23,14 +24,11 @@ namespace dxvk {
     PSCount
   };
 
-  constexpr size_t DxsoMaxTempRegs      = 32;
-  constexpr size_t DxsoMaxTextureRegs   = 10;
-  constexpr size_t DxsoMaxInterfaceRegs = 16;
-  constexpr size_t DxsoMaxOperandCount  = 8;
-
   uint32_t computeResourceSlotId(
           DxsoProgramType shaderStage,
           DxsoBindingType bindingType,
           uint32_t        bindingIndex);
+
+  uint32_t RegisterLinkerSlot(DxsoSemantic semantic);
 
 }
